# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "La bonne affaire",
        "description": "",
        "contact": {
            "email": "nathalietaffot@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/annonces": {
            "get": {
                "tags": [
                    "Annonce"
                ],
                "summary": "Get all annonces",
                "description": "Get all annonces",
                "operationId": "inde",
                "responses": {
                    "201": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "example": "200"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "affichage des annonces."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "example": "400"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur lors du traitement de la demande"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "example": "500"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur de connexion"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Annonce"
                ],
                "summary": "Register",
                "description": "Register",
                "operationId": "stores",
                "requestBody": {
                    "description": "Enregistrement d'une annonce",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "votre title *",
                                        "type": "string",
                                        "format": "string",
                                        "example": "la porte"
                                    },
                                    "size": {
                                        "description": "votre size",
                                        "type": "string",
                                        "format": "string",
                                        "example": "M"
                                    },
                                    "images": {
                                        "description": "images de vos produits *",
                                        "type": "string",
                                        "format": "images",
                                        "example": "image.png"
                                    },
                                    "description": {
                                        "description": "votre description *",
                                        "type": "string",
                                        "format": "string",
                                        "example": "long texte"
                                    },
                                    "state_product": {
                                        "description": "l'etat de votre produits *",
                                        "type": "string",
                                        "format": "string",
                                        "example": "NEUF"
                                    },
                                    "color": {
                                        "description": "color de votre produit",
                                        "type": "string",
                                        "format": "string",
                                        "example": "bleu"
                                    },
                                    "localization": {
                                        "description": "votre localization *",
                                        "type": "string",
                                        "format": "string",
                                        "example": "yaounde"
                                    },
                                    "price": {
                                        "description": "votre price *",
                                        "type": "string",
                                        "format": "string",
                                        "example": "16000"
                                    },
                                    "option_discutable": {
                                        "description": "votre option *",
                                        "type": "string",
                                        "format": "string",
                                        "example": "OUI"
                                    },
                                    "date_publish": {
                                        "description": "date publication *",
                                        "type": "string",
                                        "format": "string",
                                        "example": "date_publish"
                                    },
                                    "job_type": {
                                        "description": "type de votre produit",
                                        "type": "string",
                                        "format": "string",
                                        "example": "pantalon"
                                    },
                                    "categorie_id": {
                                        "description": "nom de categorie *",
                                        "type": "integer",
                                        "format": "string",
                                        "example": "1"
                                    },
                                    "availability": {
                                        "description": "votre availability *",
                                        "type": "string",
                                        "format": "string",
                                        "example": "INDISPONIBLE"
                                    },
                                    "state_annonce": {
                                        "description": "etat de votre annonce *",
                                        "type": "string",
                                        "format": "string",
                                        "example": "VALIDATE"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "example": "200"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Annonce bien Creer."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "example": "400"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur lors du traitement de la demande"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "example": "500"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur de connexion"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/users/users": {
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "delete password",
                "description": "delete password",
                "operationId": "destroy",
                "responses": {
                    "201": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "example": "200"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "modification du mot de passe."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "example": "400"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur lors du traitement de la demande"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "example": "500"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur de connexion"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}